requirements.txt
langchain-core
langchain
python-dotenv
openai